requirements.txt
anyio==4.9.0
appdirs==1.4.4
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.4
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
h11==0.16.0
htmltools==0.6.0
httplib2==0.22.0
idna==3.10
joblib==1.5.1
kiwisolver==1.4.8
linkify-it-py==2.0.3
markdown-it-py==3.0.0
matplotlib==3.10.3
mdit-py-plugins==0.4.2
mdurl==0.1.2
narwhals==1.43.0
nltk==3.9.1
numpy==2.3.0
orjson==3.10.18
packaging==25.0
pandas==2.3.0
pillow==11.2.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
questionary==2.1.0
regex==2024.11.6
requests==2.32.4
rsa==4.9.1
shiny==1.4.0
six==1.17.0
sniffio==1.3.1
starlette==0.47.0
textblob==0.19.0
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
uc-micro-py==1.0.3
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
